Added Rpc and Proxy module.
The Proxy uses STM and queues as a buffer between the client and the server.
This is good for clients such as emacs.

This is mostly a 1-1 translation from emacs-lsp-booster. The main difference
being that lsp-booster uses channels and an atomic int as a counter to decide
when to apply backpressure the client -> server channel, lspipe uses a bounded
TQueue making the logic a bit easier to follow in lspipe.

Instead of using Result/Either to manage errors we use effectful Error which is
basically the same a checked exception, meaning the caller must handle it or
propogate it.

Lspipe uses STM to ensure safe concurrent behaviour.